Moyenne enlevée (maudit bash !)

+++ src/d2p.c
@@ -137,7 +137,7 @@ int main(int argc, char *argv[]) {
     #ifdef TEST
       // Affichage du temps
       double elapsed_time = (end - start) * 100000000;
-      printf("%d\n", (int) elapsed_time);
+      printf("%f\n", elapsed_time);
     #else
       // Affichage du résultat
       for(int i = 0; i < 2 * n; i++)
